requirements.txt
setuptools>=61.0.0
wheel
pip>=23.1

# App runtime
streamlit==1.26.0

# Use a TensorFlow version that has published wheels for your target Python version
tensorflow==2.20.0

# Use a modern numpy wheel that supports recent Python versions
numpy==1.26.4

# Use a Pillow release with binary wheels available on many platforms
pillow==9.4.0
opencv-python==4.8.0.76
scikit-learn==1.2.2
matplotlib==3.7.1
